+++ .eslintrc.js
@@ -22,6 +22,7 @@ module.exports = {
     "plugin:ssr-friendly/recommended",
     "plugin:jsx-a11y/recommended",
     "plugin:storybook/recommended",
+    "plugin:import/typescript",
   ],
   parser: "@typescript-eslint/parser",
   parserOptions: {
@@ -33,6 +34,7 @@ module.exports = {
     "sort-imports-es6-autofix",
     "ssr-friendly",
     "jsx-a11y",
+    "jest",
   ],
   rules: {
     "@typescript-eslint/adjacent-overload-signatures": "error",
@@ -74,7 +76,7 @@ module.exports = {
     "@typescript-eslint/class-name-casing": "off",
     "@typescript-eslint/consistent-type-assertions": "off",
     "@typescript-eslint/consistent-type-definitions": "off",
-    "@typescript-eslint/explicit-function-return-type": "error",
+    "@typescript-eslint/explicit-function-return-type": "off",
     "@typescript-eslint/explicit-member-accessibility": [
       "off",
       {
@@ -179,6 +181,11 @@ module.exports = {
             position: "after",
           },
           {
+            pattern: "~*/**",
+            group: "internal",
+            position: "before",
+          },
+          {
             pattern: "..",
             group: "parent",
             position: "after",
@@ -192,6 +199,19 @@ module.exports = {
         pathGroupsExcludedImportTypes: [],
       },
     ],
+    "jest/consistent-test-it": ["error", { fn: "it" }],
+    "jest/expect-expect": "error",
+    "jest/no-commented-out-tests": "error",
+    "jest/no-conditional-expect": "error",
+    "jest/no-focused-tests": "error",
+    "jest/no-disabled-tests": "error",
+    "jest/no-standalone-expect": "error",
+    "jest/no-test-return-statement": "error",
+    "jest/prefer-equality-matcher": "error",
+    "jest/require-top-level-describe": "error",
+    "jest/valid-describe-callback": "error",
+    "jest/valid-expect": "error",
+    "jest/valid-title": "error",
     "jsdoc/check-alignment": "off",
     "jsdoc/check-indentation": "off",
     "jsdoc/newline-after-description": "off",
@@ -229,6 +249,14 @@ module.exports = {
     "no-underscore-dangle": "error",
     "no-unsafe-finally": "error",
     "no-unused-labels": "error",
+    "no-unused-vars": "off",
+    "@typescript-eslint/no-unused-vars": [
+      "error",
+      {
+        varsIgnorePattern: "^_",
+        argsIgnorePattern: "^_",
+      },
+    ],
     "no-var": "error",
     "object-shorthand": "error",
     "one-var": ["error", "never"],
@@ -260,6 +288,12 @@ module.exports = {
       files: ["*.stories.tsx", "*.spec.tsx", "*.docsExample.tsx"],
       rules: {
         "import/no-extraneous-dependencies": "off",
+      },
+    },
+    {
+      files: ["*.ts", "*.mts", "*.cts", "*.tsx"],
+      rules: {
+        "@typescript-eslint/explicit-function-return-type": "error",
       },
     },
   ],