+++ draft-packages/modal/KaizenDraft/Modal/Presets/ContextModal.spec.tsx
@@ -1,9 +1,7 @@
 import React from "react"
-import { cleanup, render, fireEvent } from "@testing-library/react"
+import { render, fireEvent } from "@testing-library/react"
 import { ContextModal, ContextModalProps } from "./ContextModal"
 import "./matchMedia.mock"
-
-afterEach(cleanup)
 
 const ContextModalWrapper = (
   props: Partial<ContextModalProps>