+++ draft-packages/title-block-zen/docs/TitleBlockZen.stories.tsx
@@ -861,6 +861,30 @@ export const DefaultOnlyPrimary: StoryFn = () => (
 )
 DefaultOnlyPrimary.storyName = "Default (only primary action)"
 
+export const DefaultOnlySecondary: StoryFn = () => (
+  <OffsetPadding>
+    <TitleBlockZen
+      title="Page title"
+      surveyStatus={{ text: "Live", status: "live" }}
+      secondaryActions={SECONDARY_ACTIONS}
+      handleHamburgerClick={(): void => alert("Hamburger clicked")}
+      breadcrumb={{
+        path: "#",
+        text: "Back to home",
+        handleClick: () => alert("breadcrumb clicked!"),
+      }}
+      navigationTabs={[
+        <NavigationTab text="Label" href="#" active />,
+        <NavigationTab text="Label" href="#" />,
+        <NavigationTab text="Label" href="#" />,
+        <NavigationTab text="Label" href="#" />,
+        <NavigationTab text="Label" href="#" />,
+        <NavigationTab text="Label" href="#" />,
+      ]}
+    />
+  </OffsetPadding>
+)
+DefaultOnlySecondary.storyName = "Default (only secondary action)"
 export const DefaultWithReportSwitcher: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen