+++ packages/button/src/Button/components/GenericButton.tsx
@@ -104,6 +104,7 @@ const GenericButton = forwardRef(
         className={classNames(styles.container, {
           [styles.fullWidth]: props.fullWidth,
         })}
+        aria-live="polite"
       >
         {determineButtonRenderer()}
       </span>
@@ -157,7 +158,6 @@ const renderButton = (
     disableTabFocusAndIUnderstandTheAccessibilityImplications,
     onFocus,
     onBlur,
-    iconButton,
     ...rest
   } = props
   const customProps = getCustomProps(rest)
@@ -195,7 +195,6 @@ const renderLink = (props: Props, ref: Ref<HTMLAnchorElement>): JSX.Element => {
     newTabAndIUnderstandTheAccessibilityImplications,
     onFocus,
     onBlur,
-    iconButton,
     ...rest
   } = props
   const customProps = getCustomProps(rest)