+++ packages/components/src/index.ts
@@ -1,2 +1,4 @@
-// Components inside this package
 export * from "./KaizenProvider"
+export * from "./Filter"
+export * from "./FilterButton"
+export * from "./FilterSelect"