+++ packages/components/src/types/OverrideClassName.ts
@@ -1,0 +1,2 @@
+export type OverrideClassName<T extends Record<string, any>> =
+  | Omit<T, "className"> & { classNameOverride?: string }