+++ packages/design-tokens/src/__tests__/sass-tokens.spec.ts
@@ -1,3 +1,4 @@
+/* eslint-disable jest/consistent-test-it */
 import fs from "fs"
 import path from "path"
 import PostCSS, { Declaration } from "postcss"