+++ packages/notification/src/GlobalNotification.spec.tsx
@@ -1,25 +1,25 @@
 import React from "react"
-import { cleanup, render } from "@testing-library/react"
+import { render } from "@testing-library/react"
 import { GlobalNotification } from "./GlobalNotification"
 
-afterEach(cleanup)
+describe("<GlobalNotification />", () => {
+  it("renders the positive notification correctly", () => {
+    const { container } = render(
+      <GlobalNotification type="positive">
+        This is my positive notification
+      </GlobalNotification>
+    )
 
-test("The basic notification renders correctly", () => {
-  const { container } = render(
-    <GlobalNotification type="positive">
-      This is my positive notification
-    </GlobalNotification>
-  )
+    expect(container.firstChild).toMatchSnapshot()
+  })
 
-  expect(container.firstChild).toMatchSnapshot()
-})
+  it("renders the informative notification correctly", () => {
+    const { container } = render(
+      <GlobalNotification type="informative">
+        This is my <em>informational</em> notification
+      </GlobalNotification>
+    )
 
-test("You can change the type", () => {
-  const { container } = render(
-    <GlobalNotification type="informative">
-      This is my <em>informational</em> notification
-    </GlobalNotification>
-  )
-
-  expect(container.firstChild).toMatchSnapshot()
+    expect(container.firstChild).toMatchSnapshot()
+  })
 })