+++ draft-packages/avatar/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.8.34](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-avatar@2.8.33...@kaizen/draft-avatar@2.8.34) (2023-03-29)
+
+**Note:** Version bump only for package @kaizen/draft-avatar
+
+
+
+
+
+## [2.8.33](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-avatar@2.8.32...@kaizen/draft-avatar@2.8.33) (2023-03-29)
+
+**Note:** Version bump only for package @kaizen/draft-avatar
+
+
+
+
+
+## [2.8.32](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-avatar@2.8.31...@kaizen/draft-avatar@2.8.32) (2023-03-23)
+
+**Note:** Version bump only for package @kaizen/draft-avatar
+
+
+
+
+
+## [2.8.31](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-avatar@2.8.30...@kaizen/draft-avatar@2.8.31) (2023-03-22)
+
+**Note:** Version bump only for package @kaizen/draft-avatar
+
+
+
+
+
 ## [2.8.30](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-avatar@2.8.29...@kaizen/draft-avatar@2.8.30) (2023-03-14)
 
 **Note:** Version bump only for package @kaizen/draft-avatar