+++ draft-packages/card/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-card@3.2.7...@kaizen/draft-card@3.2.8) (2023-03-23)
+
+**Note:** Version bump only for package @kaizen/draft-card
+
+
+
+
+
 ## [3.2.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-card@3.2.6...@kaizen/draft-card@3.2.7) (2023-03-14)
 
 **Note:** Version bump only for package @kaizen/draft-card