+++ draft-packages/guidance-block/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.1.26](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@6.1.25...@kaizen/draft-guidance-block@6.1.26) (2023-03-30)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [6.1.25](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@6.1.24...@kaizen/draft-guidance-block@6.1.25) (2023-03-29)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [6.1.24](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@6.1.23...@kaizen/draft-guidance-block@6.1.24) (2023-03-29)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [6.1.23](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@6.1.22...@kaizen/draft-guidance-block@6.1.23) (2023-03-23)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [6.1.22](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@6.1.21...@kaizen/draft-guidance-block@6.1.22) (2023-03-22)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [6.1.21](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@6.1.20...@kaizen/draft-guidance-block@6.1.21) (2023-03-22)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
 ## [6.1.20](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@6.1.19...@kaizen/draft-guidance-block@6.1.20) (2023-03-14)
 
 **Note:** Version bump only for package @kaizen/draft-guidance-block