+++ draft-packages/illustration/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.0.25](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-illustration@6.0.24...@kaizen/draft-illustration@6.0.25) (2023-03-29)
+
+**Note:** Version bump only for package @kaizen/draft-illustration
+
+
+
+
+
+## [6.0.24](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-illustration@6.0.23...@kaizen/draft-illustration@6.0.24) (2023-03-29)
+
+**Note:** Version bump only for package @kaizen/draft-illustration
+
+
+
+
+
+## [6.0.23](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-illustration@6.0.22...@kaizen/draft-illustration@6.0.23) (2023-03-23)
+
+**Note:** Version bump only for package @kaizen/draft-illustration
+
+
+
+
+
+## [6.0.22](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-illustration@6.0.21...@kaizen/draft-illustration@6.0.22) (2023-03-22)
+
+**Note:** Version bump only for package @kaizen/draft-illustration
+
+
+
+
+
 ## [6.0.21](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-illustration@6.0.20...@kaizen/draft-illustration@6.0.21) (2023-03-14)
 
 **Note:** Version bump only for package @kaizen/draft-illustration