+++ packages/brand-moment/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.13.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.12.74...@kaizen/brand-moment@1.13.0) (2023-03-30)
+
+
+### Features
+
+* 71-working-button-state-needs-to-be-announced-with-a-live-region ([#3406](https://github.com/cultureamp/kaizen-design-system/issues/3406)) ([ae55e7a](https://github.com/cultureamp/kaizen-design-system/commit/ae55e7acd8ef6d9d0797a5f9adc9e206a9995f66))
+
+
+
+
+
+## [1.12.74](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.12.73...@kaizen/brand-moment@1.12.74) (2023-03-29)
+
+**Note:** Version bump only for package @kaizen/brand-moment
+
+
+
+
+
+## [1.12.73](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.12.72...@kaizen/brand-moment@1.12.73) (2023-03-29)
+
+**Note:** Version bump only for package @kaizen/brand-moment
+
+
+
+
+
+## [1.12.72](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.12.71...@kaizen/brand-moment@1.12.72) (2023-03-23)
+
+**Note:** Version bump only for package @kaizen/brand-moment
+
+
+
+
+
+## [1.12.71](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.12.70...@kaizen/brand-moment@1.12.71) (2023-03-23)
+
+**Note:** Version bump only for package @kaizen/brand-moment
+
+
+
+
+
+## [1.12.70](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.12.69...@kaizen/brand-moment@1.12.70) (2023-03-22)
+
+**Note:** Version bump only for package @kaizen/brand-moment
+
+
+
+
+
 ## [1.12.69](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/brand-moment@1.12.68...@kaizen/brand-moment@1.12.69) (2023-03-14)
 
 **Note:** Version bump only for package @kaizen/brand-moment