+++ packages/select/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.13.20](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/select@6.13.19...@kaizen/select@6.13.20) (2023-03-30)
+
+**Note:** Version bump only for package @kaizen/select
+
+
+
+
+
+## [6.13.19](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/select@6.13.18...@kaizen/select@6.13.19) (2023-03-30)
+
+**Note:** Version bump only for package @kaizen/select
+
+
+
+
+
+## [6.13.18](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/select@6.13.17...@kaizen/select@6.13.18) (2023-03-29)
+
+**Note:** Version bump only for package @kaizen/select
+
+
+
+
+
+## [6.13.17](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/select@6.13.16...@kaizen/select@6.13.17) (2023-03-29)
+
+**Note:** Version bump only for package @kaizen/select
+
+
+
+
+
+## [6.13.16](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/select@6.13.15...@kaizen/select@6.13.16) (2023-03-23)
+
+**Note:** Version bump only for package @kaizen/select
+
+
+
+
+
+## [6.13.15](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/select@6.13.14...@kaizen/select@6.13.15) (2023-03-23)
+
+**Note:** Version bump only for package @kaizen/select
+
+
+
+
+
+## [6.13.14](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/select@6.13.13...@kaizen/select@6.13.14) (2023-03-22)
+
+**Note:** Version bump only for package @kaizen/select
+
+
+
+
+
+## [6.13.13](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/select@6.13.12...@kaizen/select@6.13.13) (2023-03-22)
+
+**Note:** Version bump only for package @kaizen/select
+
+
+
+
+
 ## [6.13.12](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/select@6.13.11...@kaizen/select@6.13.12) (2023-03-14)
 
 **Note:** Version bump only for package @kaizen/select