+++ packages/tabs/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.40](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tabs@1.5.39...@kaizen/tabs@1.5.40) (2023-03-30)
+
+**Note:** Version bump only for package @kaizen/tabs
+
+
+
+
+
+## [1.5.39](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tabs@1.5.38...@kaizen/tabs@1.5.39) (2023-03-29)
+
+**Note:** Version bump only for package @kaizen/tabs
+
+
+
+
+
+## [1.5.38](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tabs@1.5.37...@kaizen/tabs@1.5.38) (2023-03-29)
+
+**Note:** Version bump only for package @kaizen/tabs
+
+
+
+
+
+## [1.5.37](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tabs@1.5.36...@kaizen/tabs@1.5.37) (2023-03-23)
+
+**Note:** Version bump only for package @kaizen/tabs
+
+
+
+
+
+## [1.5.36](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tabs@1.5.35...@kaizen/tabs@1.5.36) (2023-03-22)
+
+**Note:** Version bump only for package @kaizen/tabs
+
+
+
+
+
 ## [1.5.35](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tabs@1.5.34...@kaizen/tabs@1.5.35) (2023-03-14)
 
 **Note:** Version bump only for package @kaizen/tabs